@@ -1,6 +1,6 @@
 module github.com/rivik/go-aux
 
-go 1.15
+go 1.16
 
 require (
 	github.com/ClickHouse/clickhouse-go v1.4.3
